Fix performance, concurrency and persistence defects
tst: stop the prefix walk at k results (1155us -> 1.8us at 100k tokens)
freq_store: exclusive lock on bump(), persistence moved to a background
  writer, tokens with spaces no longer dropped on load
stack: state-swapping undo/redo so redo moves the document forward
lru: add the missing destructor, delete copy operations
phrase_store: allow '|' inside snippets, which aborted the editor on load
editor: wire up Ranker and KMP, bind Ctrl+S, Ctrl+Z, Ctrl+Y and Esc
tests: regression suite pinning each fix, plus make test/cli targets

+++ src/minheap.cpp
@@ -13,10 +13,10 @@ void MinHeap::heapifyDown(int index) {
     int smallest=index;
     int left=leftChild(index);
     int right=rightChild(index);
-    if(left<heap.size() && heap[left].first<heap[smallest].first){
+    if(left<(int)heap.size() && heap[left].first<heap[smallest].first){
         smallest=left;
     }
-    if(right<heap.size() && heap[right].first<heap[smallest].first){
+    if(right<(int)heap.size() && heap[right].first<heap[smallest].first){
         smallest=right;
     }
     if(smallest!=index){
@@ -25,7 +25,7 @@ void MinHeap::heapifyDown(int index) {
     }
 }
 void MinHeap::insert(double score,const string& word) {
-    if(heap.size()<maxSize){
+    if((int)heap.size()<maxSize){
         heap.push_back({score,word});
         heapifyUp(heap.size()-1);
     }
